break out deploying pool registry and deploying pools
